*** empty log message ***

+++ src_tests/ServidorAplicacao/Servicos/manager/ReadDegreeServiceTest.java
@@ -18,32 +18,20 @@ public class ReadDegreeServiceTest extends TestCaseManagerReadServices {
 		super(testName);
 	 }
 
-	/* (non-Javadoc)
-	 * @see ServidorAplicacao.Servicos.TestCaseNeedAuthorizationServices#getNameOfServiceToBeTested()
-	 */
 	 protected String getNameOfServiceToBeTested() {
 		return "ReadDegree";
 	 }
        
-	/* (non-Javadoc)
-	 * @see ServidorAplicacao.Servicos.TestCaseReadServices#getArgumentsOfServiceToBeTestedUnsuccessfuly()
-	 */
 	protected Object[] getArgumentsOfServiceToBeTestedUnsuccessfuly() {
-		Object[] args = { new Integer(5) };
+		Object[] args = { new Integer(100) };
 		return args;
 	}	
 
-	/* (non-Javadoc)
-	 * @see ServidorAplicacao.Servicos.TestCaseReadServices#getArgumentsOfServiceToBeTestedSuccessfuly()
-	 */
 	protected Object[] getArgumentsOfServiceToBeTestedSuccessfuly() {
 		Object[] args = { new Integer(9) };
 		return args;
 	}
-        
-   /* (non-Javadoc)
-	* @see ServidorAplicacao.Servicos.TestCaseReadServices#getObjectToCompare()
-	*/	 
+        	 
 	protected Object getObjectToCompare() {			
 		return new InfoDegree("MEEC", "Engenharia Electrotecnica e de Computadores");
 	}
